patch from Kahr:
This patch changes the name of minecraft.original.jar to
minecraft-<version>.jar.  The version number is parsed out of the jar via
regex.  So instead of overwriting 1.2.6 with a backup of 1.2.5, it will
look for minecraft-1.2.6.jar, not find it, and correctly conclude that
this is a new version.

Another thing it does is try to detect when custom lava and water would
break a new version of minecraft.jar.  If the obfuscated class names that
ClassFinder maps StillLava, etc., to are not found in the newcode/
directory, the checkboxes are automatically unchecked and greyed out with
a short tooltip explanation.